fix: reimplement alignment stripes on top of signed ranges
This reimplements `Stripe` struct to contain a `Range<isize>` instead of a pair of `usize`s. This way it's easier to catch integer bugs and we get all existing checks and fixed of the `Range` struct.

+++ packages_rs/nextclade/src/align/align.rs
@@ -74,7 +74,7 @@ pub fn align_aa(
   gap_open_close: &[i32],
   params: &AlignPairwiseParams,
   band_width: usize,
-  mean_shift: i32,
+  mean_shift: isize,
 ) -> AlignmentOutput<Aa> {
   let stripes = simple_stripes(mean_shift, band_width, ref_seq.len(), qry_seq.len());
